- request url includes /api for packages controller - small refactoring - fixed delete bug

+++ src/main/java/com/synergo/deliverybe/services/PackageService.java
@@ -63,7 +63,8 @@ public class PackageService {
     }
 
     public void deleteById(Integer id) {
-        packageRepo.deleteById(id);
+        if(packageRepo.findById(id).isPresent())
+            packageRepo.deleteById(id);
     }
 
     public Optional<Package> updatePackage(Package pack, Integer id) {
